+++ src/Migrations/Version20260706085400.php
@@ -20,7 +20,7 @@ final class Version20260706085400 extends EntityManagerMigration
 
     public function up(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE authorization_group_resource_members TO '.ResourceGroupMember::TABLE_NAME);
+        $this->addSql('ALTER TABLE authorization_group_resource_members RENAME TO '.ResourceGroupMember::TABLE_NAME);
         $this->addSql('ALTER TABLE '.AuthorizationResource::TABLE_NAME.' ADD COLUMN '.AuthorizationResource::RESOURCE_TYPE_COLUMN.' BOOLEAN NOT NULL DEFAULT FALSE');
     }
 }